# web/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Lunches API",
        "description": "REST API of Lunches platform e-commerce solution",
        "contact": {
            "name": "Lunches API Team",
            "url": "https://lunches.com.ua",
            "email": "support@lunches.com.ua"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List all orders",
                "description": "Retrieves the list of orders by filters",
                "operationId": "getOrdersAction",
                "parameters": [
                    {
                        "name": "shipmentDate",
                        "in": "query",
                        "description": "Filter orders which will be shipped on specified date",
                        "type": "string",
                        "format": "date"
                    },
                    {
                        "$ref": "#/parameters/startDate"
                    },
                    {
                        "$ref": "#/parameters/endDate"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Orders",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Order"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Place an order",
                "description": "Creates new order",
                "operationId": "postOrdersAction",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Include here payload in Order representation",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Newly placed Order",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                }
            }
        },
        "/orders/{orderId}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Retrieve single order",
                "description": "Retrieves the details of an order by ID",
                "operationId": "getOrderAction",
                "parameters": [
                    {
                        "$ref": "#/parameters/orderId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an order",
                "description": "Updates specified Order",
                "operationId": "putOrderAction",
                "parameters": [
                    {
                        "$ref": "#/parameters/orderId"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated order",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                }
            }
        },
        "/users/{username}/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Retrieve user orders",
                "description": "Get list of user orders using filters if needed",
                "operationId": "getUserOrdersAction",
                "parameters": [
                    {
                        "$ref": "#/parameters/username"
                    },
                    {
                        "$ref": "#/parameters/startDate"
                    },
                    {
                        "$ref": "#/parameters/endDate"
                    },
                    {
                        "name": "paid",
                        "in": "query",
                        "description": "Whether filter out non paid orders or no",
                        "type": "boolean",
                        "default": false,
                        "enum": [
                            "0",
                            "1",
                            false,
                            true
                        ]
                    },
                    {
                        "name": "withCanceled",
                        "in": "query",
                        "description": "Whether include canceled orders or no",
                        "type": "boolean",
                        "default": false,
                        "enum": [
                            "0",
                            "1",
                            false,
                            true
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Orders",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Order"
                            }
                        }
                    }
                }
            }
        },
        "/orders/{orderId}/cancel": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel an order",
                "description": "Cancels an order",
                "operationId": "postOrdersCancelAction",
                "parameters": [
                    {
                        "$ref": "#/parameters/orderId"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Canceled Order",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                }
            }
        },
        "/orders/{orderId}/reject": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Reject an order",
                "description": "Reject an order",
                "operationId": "postOrdersRejectAction",
                "parameters": [
                    {
                        "$ref": "#/parameters/orderId"
                    },
                    {
                        "$ref": "#/parameters/accessToken"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Rejected Order",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Dish": {
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "description": "Allowed three values: meat, garnish, salad",
                    "type": "string",
                    "enum": [
                        "meat",
                        "garnish",
                        "salad"
                    ]
                },
                "created": {
                    "description": "$created",
                    "type": "string",
                    "format": "date-time"
                },
                "updated": {
                    "description": "$updated",
                    "type": "string",
                    "format": "date-time"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DishImage"
                    }
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Ingredient"
                    }
                }
            },
            "type": "object"
        },
        "DishImage": {
            "required": [
                "image",
                "dish"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "readOnly": true
                },
                "image": {
                    "$ref": "#/definitions/Image"
                },
                "isCover": {
                    "type": "boolean"
                },
                "dish": {
                    "$ref": "#/definitions/Dish"
                },
                "created": {
                    "description": "$created",
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "updated": {
                    "description": "$updated",
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                }
            },
            "type": "object"
        },
        "Image": {
            "required": [
                "id",
                "url",
                "format",
                "width",
                "height"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "created": {
                    "description": "$created",
                    "type": "string",
                    "format": "date-time"
                },
                "updated": {
                    "description": "$updated",
                    "type": "string",
                    "format": "date-time"
                }
            },
            "type": "object"
        },
        "Ingredient": {
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "created": {
                    "description": "$created",
                    "type": "string",
                    "format": "date-time"
                },
                "updated": {
                    "description": "$updated",
                    "type": "string",
                    "format": "date-time"
                },
                "dish": {
                    "$ref": "#/definitions/Dish"
                }
            },
            "type": "object"
        },
        "LineItem": {
            "required": [
                "size",
                "product"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "size": {
                    "description": "Size of the dish portion",
                    "type": "string",
                    "enum": [
                        "small",
                        "medium",
                        "big"
                    ]
                },
                "dish": {
                    "$ref": "#/definitions/Dish"
                },
                "order": {
                    "$ref": "#/definitions/Order"
                }
            },
            "type": "object"
        },
...
        "Order": {
            "required": [
                "user",
                "address",
                "shipmentDate",
                "lineItems"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "readOnly": true
                },
                "orderNumber": {
                    "description": "Number of order. Starts from 1000. Generated automatically, no need to specify custom order number as it will be ignored",
                    "type": "string",
                    "readOnly": true
                },
                "user": {
                    "description": "User who creates an Order. When you create an order, specify only **userId**",
                    "$ref": "#/definitions/User"
                },
                "address": {
                    "description": "The shipping address for the order. Can be omitted, in this case user default address will be used",
                    "type": "string"
                },
                "created": {
                    "description": "Date and time when order was created",
                    "readOnly": true,
                    "$ref": "#/definitions/CreatedOrder"
                },
                "canceled": {
                    "readOnly": true,
                    "$ref": "#/definitions/CanceledOrder"
                },
                "rejected": {
                    "readOnly": true,
                    "$ref": "#/definitions/RejectedOrder"
                },
                "delivered": {
                    "readOnly": true,
                    "$ref": "#/definitions/DeliveredOrder"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "created",
                        "inProgress",
                        "canceled",
                        "rejected",
                        "delivered",
                        "closed"
                    ],
                    "readOnly": true
                },
                "shipmentDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "price": {
                    "description": "Order price. Read only. Calculates automatically and based on order line items cost sum, delivery cost and taxes.\nIf user orders all dishes of menu, orders costs lower then ordering of dishes separately",
                    "type": "number",
                    "format": "float",
                    "readOnly": true
                },
                "paid": {
                    "description": "Order payment status with boolean value",
                    "type": "boolean",
                    "readOnly": true
                },
                "lineItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/LineItem"
                    }
                }
            },
            "type": "object"
        },
...
        "User": {
            "required": [
                "username",
                "address"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "readOnly": true
                },
                "username": {
                    "type": "string"
                },
                "balance": {
                    "description": "Current user balance. It updates automatically with every completed transaction",
                    "type": "number",
                    "format": "float",
                    "readOnly": true
                },
                "credit": {
                    "description": "User can have credit money. It increases when user have insufficient funds to pay for an order, and decreases when order is paid successfully",
                    "type": "number",
                    "format": "float",
                    "readOnly": true
                },
                "clientId": {
                    "description": "Auto generated human readable user ID",
                    "type": "integer",
                    "readOnly": true
                },
                "address": {
                    "description": "Default user address. It will be used, when there were no address specified for an Order",
                    "type": "string"
                },
                "created": {
                    "description": "$created",
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                }
            },
            "type": "object"
        },
        "CanceledOrder": {
            "properties": {
                "at": {
                    "description": "Date and time when order was canceled",
                    "type": "string",
                    "format": "date-time"
                },
                "reason": {
                    "description": "The reason why the order was canceled",
                    "type": "string"
                }
            }
        },
        "CreatedOrder": {
            "properties": {
                "at": {
                    "description": "Date and time when order was created",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "DeliveredOrder": {
            "properties": {
                "at": {
                    "description": "Date and time when order was delivered",
                    "type": "string",
                    "format": "date-time"
                },
                "carrier": {
                    "description": "Delivery service which carried out the order",
                    "type": "string"
                }
            }
        },
        "RejectedOrder": {
            "properties": {
                "at": {
                    "description": "Date and time when order was rejected",
                    "type": "string",
                    "format": "date-time"
                },
                "reason": {
                    "description": "The reason why the order was rejected",
                    "type": "string"
                }
            }
        }
    },
    "parameters": {
        "accessToken": {
            "name": "accessToken",
            "in": "query",
            "description": "Authentication token to access restricted resources",
            "required": true,
            "type": "string"
        },
...
        "startDate": {
            "name": "startDate",
            "in": "query",
            "description": "Filter data which _date_ value greater than **startDate**",
            "type": "string",
            "format": "date"
        },
        "endDate": {
            "name": "endDate",
            "in": "query",
            "description": "Filter data which _date_ value less than **endDate**",
            "type": "string",
            "format": "date"
        },
        "orderId": {
            "name": "orderId",
            "in": "path",
            "description": "ID of Order",
            "required": true,
            "type": "integer"
        },
        "username": {
            "name": "username",
            "in": "path",
            "description": "User name",
            "required": true,
            "type": "string"
        },
...
    },
...
}